leds: avoid spurious reappearance of LEDS during cassette reading.
When reading data from cassette the OS briefly turns off the cassette
motor relay between blocks.  When LEDs were to set to "when changed" or
"when changed or transient" this caused the LED bar to re-appear then
fade out again.  Accoring to a discussion on the forums, the LED
momentarily going out is not percepible.

https://stardot.org.uk/forums/viewtopic.php?f=4&p=280528

This comment therefore adds a delay between a 0 being written to the bit
in the Serial ULA and turning of the B-Em cassette LED.  This avoids the
bar popping up when there is nothing of interest on it.