keyboard: add support for PC/XT layout (needs pathed Allegro).
The IBM Model F keyboard as used on the PC and PC/XT does not have
dedicated cursor keys or dedicated keys for Home, End, Insert, Delete,
Pgup, PgDown but instead has a dual purpose keypad which can be numeric
or fullfil these other functions depend on the state of numlock.

Allegro always returns a code like ALLEGRO_KEY_PAD_0 so this adds code
to check the state of numlock and map these to the codes for the
separate keys if numlock is off.  This also needs a patched version of
Allegro which send the modifier status with key down/up events as well
as key char events.