keyboard: make keypad mode optional.
This keypad mode enables a pre-IBM AT keyboard, for example Model F, to
be used in B-Em.  When the Numlock is off the keypad is treated as if it
was sending the standard cursor and editing keys.  When numlock is on it
sends the BBC Master keypad codes.

Note at the moment this needs the use of a patched version of Allegro
that sends the modifier state even with key up/down events (not just key
char events).