keyboard: generate cursor keys in logical mode with numlock off.
This changes the behaviour of the PC numeric keypad.  This is affected
by three things - logical or physical mapping, the PC/XT keypad mode and
the state or numlock.  Prior to this commit, in logical mode with keypad
mode off and numlock off the keypad did not do anything at all.  This
commit changes that to generating cursor keys by having logical mode
always behave as if keypad mode is active.