First step to providing a DFS-mode.
This commit allows the file lookups in general and the *DIR and *LIB
commands to be swapped so that VDFS behaves a bit more like DFS rather
than ADFS.  This mode is selected when FSCLAIM +D (claim DFS) is on and
VDFS is selected via *DISC or via the equivalent ROM service call with
the DFS filing system number.

This does not change the format of the *CAT output or how the current
directory or library is reported.