vdfs: add directory attributes and support for *OPT4/*TITLE.
This adds the storage of directory attributes via a .inf file similar to
file attributes with the same capability to have a custom Acorn name for
a host directory.

Unlike the .inf file for a file, the one for a directory stores the boot
setting (set via *OPT 4) and the directory title, set via *TITLE.

As yet, there is no way to query these new values but the boot option is
now fully implemented at boot time.