vdfs: clean-up file directory scanning and fix boot option.
This fixes a bug with the boot option being uninitialsed if *TITLE was
run for a directory before *CAT.  In getting to the bottom of this it
was apparent trying to read the .inf file before discovering if an
object is file or directory doesn't work as the tail of the .inf line is
different between the two.  This commit therefore also reverses the
order and gets the host filing system attributes first and then parses
the .inf file.