vdfs: filename length/other lookup errors.
Prior to this commit the find_entry function would return NULL when a
name lookup failed and, if reported to the user as an error, this was
always reported as "Not found"

This commit allows find_entry to return the error that should be
reported and at same time introduces checking of the length of filename
components rather than just using those characters before the limit.  It
also changes the error message when a filename is used in a pathname as
if it were a directory name.