6502: Fix bug in paging 12K overlay RAM on B+
The new write_romsel routine (from Integra B support) internally uses
flags for various different parts of the overlay RAM that can be paged
in on a B+, Master and the Integra B.  One of these, ramk8k was being
set from an AND operation but then the value was written to RAMBank
which later on sets vis20k which is expected to be only 0 or 1 as it
indexes the memstat array.  This change converts other non-zero values
(0x80 in this case) to 1 before storing in RAMBank.