6502: fix debugger showing ROM code when executing from RAM.
In order to implement sideways ROM specific breakpoints a scheme is used
in which the ROM number is encoded in the upper 8 bits of a 32 bit
address given to the debugger and a syntax specific to this 6502 (as
opposed to a Tube 6502) for entering addresses in that format.

This also enables debugger commands such as 'd' and 'm' to display data
from a ROM other than the one currently selected.

This does not fully encode the state of all paged memory, i.e. it does
not include the state of the RAM overlay.  This commit causes reading
and writing memory for the benefit of the debugger to revert to the full
memory read/write implementation if the ROM being read from or written
to the current one.