6502: make NMIs consistently edge triggered.
The NMI handling was not consistent between the (NMOS) 6502 and the
65C02 and in one case relied on resetting the 'nmi' variable to achieve
edge triggering.  This is a problem because the peripheral can re-assert
it and avoid the edge trigger.  Now both CPUs consistently use
comparison of the current 'nmi' value with 'oldnmi' its immediately
preceeding state and only take the NMI on a postive transition.