econet: slower polling + idle wait timeout.
This uses otherstuff_poll which is called every 128 CPU cycles rather
that polling after every instruction.  To do this it has to re-calculate
some of the status after each register write and after reading the
reveive FIFO but this will still be less often than before.

The idle wait timeout is to address the perception that received packets
were not responded to by the BBC DNFS code because they arrived so
quickly (or were already in the OS buffer) that the BBC micro did not
have time to set-up the receiever's packet filter.