windows: Avoid setting compiler in Makefile.win
The Makefile.win makefile is used both from makebem.bat for native
compilation under MingW and from mkcross.sh for cross-compilation on
Linux but the compilers are typically named differently in these two
cases.

This commit removes defining the compiler from Makefile.win and relies
on the calling shell script to set for the task in hand.