vdfs: fix bug for tube operation in service call &FE.
This fixes a bug introduced by the merging the boot branch.  The message
from the tube processor is printed by the ROM installing the tube host
code in response to service call &FE.  The version of the handler for
this that also prints the boot logo was returning with A=&FE, i.e. the
call unclaimed.  If a lower numbered ROM also responded to this call it
would re-attempt reading the start up message over the tube and hang.

This also adds a missed suppression of the normal OS start-up message.