vdfs: add the ability to transle line endings (CR<>LF) on the fly.
This introduces a flag stored in the file attributes in the .inf file
that indicates both that the file is considered to be a text file and
that line endings should be translated such that files that are stored
on the host filesystem with Unix line endings, i.e. LF-only are
translated on reading to Acorn-style CR-only line endings on read and CR
is translated back to LF on write.