Issue#171: fix joystick direction.
Through the issue raised refers to the compact the issue appears to be
that there should be a negation between the natural joystick directions
from Allego and the ADC conversion on the BBC B to match the convention
used for real analogue joysticks.  The lack of this negation led to the
axes being scaled by -1 in the joymap which then caused the thesholding
in the compact joystick implementation to do the wrong thing.