6502tube: fix instruction timing on tube CMOS processor.
See: https://stardot.org.uk/forums/viewtopic.php?f=55&t=25298
and https://github.com/dp111/6502Timin

Note the timing is untested.  This is simply a copy of the changes made
to the CMOS processor in the Master which was tested.  The functionailty
has been regression tested by the Dormann tests.