Re-work WE RAM/ROM to fit with B-Em memory scheme.
This changes the support for the Watford Electronics ROM/RAM board from
intercepting every write to memory to using the existing general
structure of B-Em's memory management, i.e. memlook and memstat.

This introduces a new memory type, 3, in memstat to add to the
existing ones to make the current set:
0 - I/O
1 - RAM
2 - ROM
3 - Split R/W ROM (watford only at the moment).