vdfs: use ROM/RAM split rather than temporary S/W RAM.
The support of the W/E RAM/ROM may impact the previous technique VDFS
used to get some temporary workspace, i.e. to make its own ROM bank
temporarily a RAM bank, i.e. writable.  Also this way of working has the
disadvantage that pressing break while the bank was writable would leave
it writable.

The new approach is to implement banks which can be split ROM/RAM so the
first part of the bank is ROM and the end of the bank is RAM.  VDFS now
uses this instead and sets the split point during initialisation.  That
way its code is protected from accidental overwriting while the
workspace is always writable.