Fix Issue#193.
This bug appears to be because of change in behaviour in Allegro in that
it started, immediately after setting the fullscreen flag on the
display, returning the old window size for al_get_display_width and
al_get_display_height.  The correct values were being provided later via
a resize event.  This commit removes the calculations previously done at
the time of the resize and moves them to the handler for the resize
event.