Issue#193: Make full screen work again on Windows.
Allegro is not doing a good job at hiding differences here.

On Linux, the accurate display size is given in a resize event while the
values returned by al_get_display_width and al_get_display_height do not
change.  On Windows, no such event is sent but al_get_display_width and
al_get_display_height return the values for the full screen size
immediately after the full screen flag is set on the display.  This also
means on Windows the previous size has to remembered and restored, on
Linux this is not necessary.

Additionally, Windows requires the bodge of setting the full screen
flag, resetting it, then setting a second time!