Issue#200: DFS lock-up, disc spindown and FDC timeout.
This resolves an issue in which DFS could get stuck polling the FDC for
a status change that never happenned.  This was caused in turn because
for some order of FDC commands and seek distances, the disc spindown
counter could expire before the FDC counter for the seek operation.
This in turn was due to the disc spindown counter not being reset at the
start of a command.