config: treat failure to load config as a fatal error.
See: https://stardot.org.uk/forums/viewtopic.php?p=404842#p404842

B-Em tried to carry on without the config file but this was buggy and,
as the models are defined in the config file, ultimately would be
fruitless.  This commit makes failure to load the config file a fatal
error with a clean exit (with non-zero return to the OS).