vdfs: only copy "clean" characters between Acorn and host filenames.
This excludes control characters and characters with bit 7 set from
those copied between the Acorn filenames and host filenames.  With
regarrd to the filenames with bit 7 set, this is because we don't know
if the host will handle them in a way that is consistent with the guest,
i.e. the host may want those charcters to be part of UTF-8 encoding.