Issue#214: Scale Music 5000 speed with rest of emulator.
The Music5000 audio output was previously limiting the speed of the
emulator itself when trying to select speeds greater than 100%.

This commit changes the sampling rate declared to the host for the Music
5000 to be the right multiple of the real sampling rate for the emulator
to run at that speed.  For example, if we choose 200% for the emulator
speed we tell Allegro the sampling rate for the Music 5000 voice is
93750Hz, twice the real Music 5000 sample rate of 46875Hz.

This will also change the pitch of the Music 5000 output at speeds other
than 100%, rather like varying the speed of a tape recorder.