video: work-around failure to scale to full-screen on Linux.
See Allegro Issue#1500 (https://github.com/liballeg/allegro5/issues/1500)

Immediately after setting the full-screen flag, Allegro still reports
the display size as being the former size of the window, not the
full-screen size, for several frames.  This commit adds code to respond
to the change when it does eventually happen.