tube: fix load/save state for the new Tube ULA structure.
The Tube ULA structure has changed size due to the new latches added to
implement read empty behaviour and because some integer counters have
been reduced to 8-bit.  This implements a scheme by which the new
structure can be saved as part of savestate and, on loading, old and new
versions of the structure can be loaded back in.