autoboot: revert to simulating <Shift> for autoboot.
At some point the implementation of autoboot had been changed so that,
instead of the simulating the shift key being pressed, it altered the
start up options to invert the boot flag, i.e. so the OS would tell the
filing system to boot if the shift key was NOT held down.

Unfortunately, this is not compatible with setting autoboot before a
sofware initiated soft break on a Master because, in that case, the OS
does not re-read the startup options from CMOS so the boot flag cannot
be overridden.

This change revers to simulating the depression of the shift key.