mmb: fix DIN command and implement the DOP command.
This commit adds the DOP command however, both the DOP and DIN commands
are specified as working on the current drive if the drive number is not
specified.  The DIN command did not previously do this but defaulted to
drive zero if a drive was not specified because, not being a filing
system like MMFS, the MMB module doesn't have a current drive.  This
commit includes code to check the active filing system and, if it is
DFS, ask DFS for the current drive (via OSGBPB) to use a a better
default.