mmb: fix offset for B-sides.
Because the SDF module does not have the concept of completely different
discs in the A and B sides of each drive, we bodge this by declaring the
loaded disc as being double-sided with sequential sides.  That means the
offset for the disc image for Side B has the size of an 80t disc added
to it by the SDF code so we must subract the same to make it work.