Fix Issue#235.
This fixes Issue#235, on Linux, which is where the file locking issue
exists, and for the SDF module only (SSD, DSD etc), by checking if the
same image file is being loaded into both drives and rejecting the
second load attempt.

As part of the changes, the various disc load functions are changed
to return a status code and, in the event the disc is not loaded,
the filename isn't left on the eject menu.