Issue#226: Add a fully logical keyboard mode.
The existing logical keyboard mode is actually a hybrid mode that
assumes the letter keys on an English keyboard only generate letters and
not symbols.  This is not true of the German keyboard.

This commit implements a fully logical keyboard mode and renames the
previous one Hybrid mode.  In the fully logical mode, instead of
consulting the table allegro2bbclogical to decide whether to map the key
as physical or logical, the decision is based on whether a valid
character value is received as the unicode translation of the key.  If a
viable unicode translation is received, it is always used.