Add paste via simulated keystrokes.
For reasons of speed, the "Paste as if typed" action intercepted the
OS read from buffer routine so it would read the pasted string as if
it had come from the keyboard buffer without doing anything with the
emulated keyboard.

This does not work with OS 0.1 so this commit adds a second option to
simulate actual keypresses for the pasted in text.  This includes a
new menu item in the Edit menu, a new debugger command 'pastek' and
ad new command line option 'pastek'.

The OS intercept method remains available.