Issue#244: make FDC to disc seek interface relative.
Each FDC, i8271 and WD1770 has a track register.

In the case of the WD1770 there is only one, which is therefore swapped
by the filing system ROM when a different drive is selected.  In the
case of the i8271 there is one per drive.

With the exception of track zero, though, the FDC does not actually know
what track the disc head is on because the interface between the FDC and
the disc is a pair of outputs to step one track at a time and an input
for track zero, and because in each case the track register can be
written from software without moving the head.

When reading sectors, a real FDC will check that the sector IDs being
read from disc have their track fields the same as the contents of the
FDC's track register.  Some software protection schemes make use of this
by seeking to one (physical) track then setting the track register to a
different (logical) track number to match the sector IDs recorded there.

Prior to this commit, a bug existing with the 1770 implementation so
that, after setting the track register from software, then issuing a
seek command for the track number already in the register (logical track
number) a seek would be sent to the disc file format to seek to that
track as if it were a physical track, causing a track number mismatch
when the sectors are read.

This commit fixes that bug by moving responsibility for tracking the
position of the disc head into the disc module (disc.c) so both FDCs now
only have their own, software accessible track register and don't try to
track where the head actually is.  That in turn means the seek API
between an FDC and the disc layer changes so, instead of seeking to an
absolute track number, there is a function to seek to track zero and
another one to seek relative to the current position by the specified
number of tracks.

Responsibility for calculating when the FDC should receive a callback
after the seek is complete is also moved to the disc module (disc.c)
from ddnoise.c
issue