Issue#244: fix "Read IDs" by limiting number of data bytes.
The i8721 and the WD1770 both have commands to scan a disc track for
the sector ID fields, called the "Read ID" command on the 8271 and the
"Read address" command on the 1770, but these differ in how many bytes
of the ID are returned.  The 1770 returns the CRC bytes and the 8271
does not.

Both are implemented in the disc layer by the readaddress function but
to make this return a different number of bytes to each FDC type would
require changes in each of the disk image file modules.

Instead, this commit has the 8271 limit the number of bytes it accepts
via the fdc_data callback.  For normal sector reads this is set to the
size of the sector, for verify operations it is set to zero and for the
"Read ID" command it is set to four, to accept the track, head, sector
number and sector length fields, but not the CRC.