Issue#244: API change, no physical track numbers except seeking.
This builds on commit ed8f9cbe871a78e9ccb575b1c0ac694fa912a070.

The FDC is not expected to know for sure what physical track the heads
are on, even though most attempt to keep track of this, because each FDC
allows its own track register to be written to.  This is now tracked at
the disc layer (disc.c) and there may also be caching in the disc image
format modules, each in response to seek commands.

This commit changes the API for those commands which are not checking
the track number written into the sector ID headers, i.e. those that
will blindly use the physical track position, not to take a track number
as an argument but to use the track number from the disc layer or an
equivalent cache.