imd: Fix bugs reading and writing various maps.
Fixes a failure to add the necessary bit to the head number in the track
header to indicate the presence of a clyinder map or head map.

Checks if variable-length sectors are really variable length and, if so,
correctly writes two-byte length values.

Fixes a failure to use the cylinder and head maps values on loading.