Issue#266: Enable fractional tube multiplier/better default for AES.
This commit changes the way the number of cycles given to the tube
processor is calculated to allow a fractional multiplier, i.e. the clock
speed of the tube processor does not need to be a multiple of 2MHz or
even a round number of MHz.

This also includes a change to the default tubes table to exploit this
feature to give a more accurate speed for the tube evaluation system.