Enable config and log files to be specified on the command line.
This allows the log and config files to be specified on the command
line.  This includes a refactored command line parser and, because the
command line is now processed before the config file, there are also
changes to the way the config is loaded not to overwrite command line
options.