A Solution to screenshots with interlace and screen flicker.
See: https://www.stardot.org.uk/forums/viewtopic.php?p=457437#p457437
and: https://www.stardot.org.uk/forums/viewtopic.php?p=458944#p458944

There were issues with flickering and briefly showing old frames and
with saving only the most recent field of an interlaces picture to a
screenshot file rather than the whole frame.

Locking a bitmap with ALLEGRO_LOCK_WRITEONLY gives permission for the
host video system to leave the bitmap in any old state as the assumption
is that we will draw over the whole of it.  In the case of interlaces
output b-em assumes that the previous field will remain in the bitmap
and only draws the latest field.

On the other hand, using ALLEGRO_LOCK_READWRITE can give very poor
performance on some systems.

This commit enables which of these flags is used to be configured as an
option on the Settings->Video menu.