vdfs: partial implementaion of Tom's .inf spec.
See: https://stardot.org.uk/forums/viewtopic.php?p=463362#p463362
and: https://github.com/stardot/inf_format

This updates the way VDFS on B-Em uses .inf files to use %-encoding as
specified in Tom Seddon's new standardisation of the .inf file format
for filenames and the title string of a directory.

This is not a completely general implementation as it does not allow an
arbitrary number of hex values and it doesn't retain name value pairs it
does not understand.